Move selection headers to include folder
