+++ rviz_common/include/rviz_common/selection/selection_manager.hpp
