+++ rviz_common/src/rviz_common/temp/default_plugins/displays/grid_display.cpp
@@ -51,7 +51,7 @@
 #include "rviz_rendering/grid.hpp"
 #include "rviz_common/properties/parse_color.hpp"
 #include "rviz_common/properties/property.hpp"
-#include "../../../selection/selection_manager.hpp"
+#include "rviz_common/selection/selection_manager.hpp"
 
 using rviz_rendering::Grid;
 using rviz_common::properties::ColorProperty;
