+++ rviz_common/src/rviz_common/temp/default_plugins/tools/move_tool.cpp
@@ -32,7 +32,7 @@
 #include "rviz_common/display_context.hpp"
 #include "../../../render_panel.hpp"
 #include "../../../viewport_mouse_event.hpp"
-#include "../../../selection/selection_manager.hpp"
+#include "rviz_common/selection/selection_manager.hpp"
 #include "rviz_common/view_controller.hpp"
 #include "../../../view_manager.hpp"
 #include "../../../load_resource.hpp"
