+++ rviz_common/src/rviz_common/view_controller.cpp
@@ -57,7 +57,7 @@
 #include "rviz_common/properties/bool_property.hpp"
 // #include "./properties/enum_property.hpp"
 #include "rviz_common/properties/float_property.hpp"
-#include "./selection/selection_manager.hpp"
+#include "rviz_common/selection/selection_manager.hpp"
 #include "./render_panel.hpp"
 
 // #include "rviz/frame_manager.h"
