+++ rviz_common/src/rviz_common/visualizer_app.cpp
@@ -43,7 +43,7 @@
 #include "rviz_common/ros_integration/init.hpp"
 #include "rviz_common/ros_integration/ok.hpp"
 
-#include "./selection/selection_manager.hpp"
+#include "rviz_common/selection/selection_manager.hpp"
 #include "./visualization_frame.hpp"
 #include "./visualization_manager.hpp"
